+++ odoo/local-src/db2_import/converter/__init__.py
@@ -1,3 +1,4 @@
 from . import common
 from . import sale
 from . import purchase
+from . import ticket